@@ -1,13 +1,13 @@
 apply plugin: 'com.android.application'
 apply plugin: 'kotlin-android'
-apply plugin: 'kotlin-android-extensions'
+
 android {
-    compileSdkVersion 28
+    compileSdkVersion 32
 
     defaultConfig {
         applicationId "com.wzjing.openglesnative"
         minSdkVersion 21
-        targetSdkVersion 28
+        targetSdkVersion 32
         versionCode 2
         versionName "1.1"
 
@@ -21,6 +21,13 @@ android {
         }
 
     }
+
+
+    buildFeatures {
+        viewBinding true
+
+    }
+
 
     buildTypes {
         debug {
@@ -39,29 +46,6 @@ android {
 
     flavorDimensions "abi"
 
-    productFlavors {
-
-        arm64 {
-            versionNameSuffix "_arm64"
-            externalNativeBuild {
-                cmake {
-                    abiFilters "arm64-v8a"
-                }
-            }
-            dimension "abi"
-        }
-
-        x86 {
-            versionNameSuffix "_x86"
-            externalNativeBuild {
-                cmake {
-                    abiFilters "x86"
-                }
-            }
-            dimension "abi"
-        }
-
-    }
 
     externalNativeBuild {
         cmake {
@@ -72,13 +56,6 @@ android {
 }
 
 dependencies {
-    implementation fileTree(dir: 'libs', include: ['*.jar'])
-
-    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
-    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-android:1.1.1'
-    implementation 'com.android.support:appcompat-v7:28.0.0'
-
-    testImplementation 'junit:junit:4.12'
-    androidTestImplementation 'com.android.support.test:runner:1.0.2'
-    androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
+    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-android:1.6.2'
+    implementation 'androidx.appcompat:appcompat:1.4.2'
 }
